+++ docs/CHANGELOG.md
@@ -10,6 +10,10 @@ nav_order: 5
 
 ## main
 
+* Add `#content?` method that indicates if content has been passed to component.
+
+    *Joel Hawksley*
+
 * Added example of a custom preview controller.
 
     *Graham Rogers*